requirements.txt
django-cors-headers
Django
django-allauth
djangorestframework
geopy
psycopg2


